Move test to integration test
